requirements.txt
# ── Core ────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.12.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0

# ── ASR (Whisper) ─────────────────────────────────────────────────────────────
transformers>=4.40.0
torch>=2.1.0

# ── RAG (Corpus + Index) ──────────────────────────────────────────────────────
PyMuPDF>=1.24.0          # fitz — fast PDF text extraction
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0         # use faiss-gpu if preferred

# ── Translation (Free LLM Alternative) ────────────────────────────────────────
deep-translator>=1.11.4

# ── TTS ────────────────────────────────────────────────────────────────────────
edge-tts>=6.1.10         # Microsoft edge-tts, no API key needed
